Factored out moderngl choice from pyqt5 window
