+++ examples/basic_scene.py
@@ -1,6 +1,7 @@
-from utils.pyqt5_moderngl import show_scene
+from utils.pyqt5 import show_scene
 
 import radiant
+from radiant.renderers.moderngl import ModernGLRenderer
 
 
 def generate_scene():
@@ -24,4 +25,4 @@ def generate_scene():
 
 
 if __name__ == "__main__":
-    show_scene(*generate_scene())
+    show_scene("basic scene", *generate_scene(), ModernGLRenderer)
